moved linalg tests to submodule
